renaming, improving runtime api
