+++ aqueduct/lib/aqueduct.dart
@@ -19,9 +19,13 @@ library aqueduct;
 export 'package:logging/logging.dart';
 export 'package:safe_config/safe_config.dart';
 
+export 'package:aqueduct/src/runtime/compiler.dart';
+
 export 'src/application/application.dart';
 export 'src/auth/auth.dart';
 export 'src/db/db.dart';
 export 'src/db/managed/relationship_type.dart';
 export 'src/http/http.dart';
 export 'src/openapi/openapi.dart';
+
+
