+++ aqueduct/lib/src/cli/scripts/openapi_builder.dart
@@ -4,7 +4,6 @@ import 'package:aqueduct/aqueduct.dart';
 import 'package:aqueduct/src/cli/command.dart';
 import 'package:aqueduct/src/cli/mixins/openapi_options.dart';
 import 'package:aqueduct/src/cli/mixins/project.dart';
-import 'package:aqueduct/src/runtime/app/app.dart';
 import 'package:isolate_executor/isolate_executor.dart';
 import 'package:runtime/shim.dart';
 import 'package:yaml/yaml.dart';
