+++ aqueduct/lib/src/db/managed/data_model.dart
@@ -1,11 +1,10 @@
 import 'package:aqueduct/src/openapi/documentable.dart';
-import 'package:aqueduct/src/runtime/orm/orm.dart';
 import 'package:aqueduct/src/utilities/reference_counting_list.dart';
 
 import 'package:aqueduct/src/db/query/query.dart';
 
 import 'package:aqueduct/src/db/managed/managed.dart';
-import 'package:aqueduct/src/runtime/runtime.dart';
+import 'package:runtime/shim.dart';
 
 /// Instances of this class contain descriptions and metadata for mapping [ManagedObject]s to database rows.
 ///
