--- aqueduct/lib/src/runtime/orm/orm.dart
@@ -1,18 +1,0 @@
-import 'package:aqueduct/src/db/managed/managed.dart';
-import 'package:aqueduct/src/runtime/runtime.dart';
-
-abstract class ManagedEntityRuntime extends DynamicRuntime {
-  @override
-  String get kind => "entity";
-
-  ManagedEntity get entity;
-  ManagedObject instanceOfImplementation({ManagedBacking backing});
-  ManagedSet setOfImplementation(Iterable<dynamic> objects);
-  void setTransientValueForKey(ManagedObject object, String key, dynamic value);
-  dynamic getTransientValueForKey(ManagedObject object, String key);
-  bool isValueInstanceOf(dynamic value);
-  bool isValueListOf(dynamic value);
-
-  dynamic dynamicAccessorImplementation(Invocation invocation, ManagedEntity entity, ManagedObject object);
-  dynamic dynamicConvertFromPrimitiveValue(ManagedPropertyDescription property, dynamic value);
-}
