+++ aqueduct/lib/src/runtime/orm_impl.dart
@@ -1,7 +1,6 @@
 import 'dart:mirrors';
 
 import 'package:aqueduct/src/db/managed/managed.dart';
-import 'package:aqueduct/src/runtime/orm/orm.dart';
 import 'package:aqueduct/src/utilities/mirror_cast.dart';
 
 class ManagedEntityRuntimeImpl extends ManagedEntityRuntime {
@@ -100,8 +99,6 @@ class ManagedEntityRuntimeImpl extends ManagedEntityRuntime {
 
     return """
 import 'package:aqueduct/src/db/managed/managed.dart';
-import 'package:aqueduct/src/runtime/orm/orm.dart';
-import 'package:runtime/runtime.dart';
 import '$originalFileUri';
 
 final instance = ManagedEntityRuntimeImpl();
