+++ aqueduct/lib/src/runtime/resource_controller/bindings.dart
@@ -3,8 +3,8 @@ import 'dart:mirrors';
 import 'package:aqueduct/src/http/request.dart';
 import 'package:aqueduct/src/http/serializable.dart';
 import 'package:aqueduct/src/openapi/documentable.dart';
-import 'package:aqueduct/src/runtime/app/resource_controller_mirror/parameter.dart';
-import 'package:aqueduct/src/runtime/app/resource_controller_mirror/utility.dart';
+import 'package:aqueduct/src/runtime/resource_controller/parameter.dart';
+import 'package:aqueduct/src/runtime/resource_controller/utility.dart';
 import 'package:aqueduct/src/utilities/mirror_cast.dart';
 import 'package:aqueduct/src/utilities/mirror_helpers.dart';
 import 'package:open_api/v3.dart';
