+++ aqueduct/lib/src/runtime/resource_controller/parameter.dart
@@ -2,7 +2,7 @@ import 'dart:mirrors';
 
 import 'package:aqueduct/src/http/http.dart';
 import 'package:aqueduct/src/http/resource_controller_bindings.dart';
-import 'package:aqueduct/src/runtime/app/resource_controller_mirror/bindings.dart';
+import 'package:aqueduct/src/runtime/resource_controller/bindings.dart';
 import 'package:aqueduct/src/utilities/mirror_helpers.dart';
 
 class BoundParameter {
