--- aqueduct/lib/src/runtime/runtime.dart
@@ -1,8 +1,0 @@
-/* These two imports needs to get yanked on AOT compilation */
-export 'package:aqueduct/src/compilers/app/application_compiler.dart';
-export 'package:aqueduct/src/compilers/orm/data_model_compiler.dart';
-
-export 'package:runtime/shim.dart';
-
-export 'app/app.dart';
-export 'orm/orm.dart';
