+++ aqueduct/test/http/compilation_errors/resource_controller_filter_nonserializable_test.dart
@@ -1,7 +1,7 @@
 import 'dart:async';
 
 import 'package:aqueduct/aqueduct.dart';
-import 'package:aqueduct/src/runtime/runtime.dart';
+import 'package:runtime/shim.dart';
 import "package:test/test.dart";
 
 void main() {
